Revert types/node update and move shared functions to utils.
