+++ package.json
@@ -15,7 +15,7 @@
   "license": "MPL-2.0",
   "devDependencies": {
     "@types/dedent": "0.7.0",
-    "@types/node": "18.0.6",
+    "@types/node": "16.11.7",
     "@vercel/ncc": "0.34.0",
     "prettier": "2.7.1",
     "typescript": "4.7.4"
